# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    description: Error response structure
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  handlers.SessionValidationResponse:
    description: Session validation response structure
    properties:
      valid:
        example: true
        type: boolean
    type: object
...
  user.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  user.User:
    properties:
      name:
        type: string
      session_id:
        type: string
      username:
        type: string
    type: object
host: localhost:5055
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A RESTful Web API for managing Kronox web resources, schedules, events,
    and bookings
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kronox API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password across multiple school
        URLs
      parameters:
      - description: School that request pertains to
        example: '"hkr"'
        in: query
        name: school
        required: true
        type: string
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully authenticated
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Invalid credentials or login failed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: User login
      tags:
      - authentication
  /auth/validate:
    get:
      consumes:
      - application/json
      description: Check if a user session is valid across multiple school URLs
      parameters:
      - description: School that request pertains to
        example: '"hkr"'
        in: query
        name: school
        required: true
        type: string
      - description: Bearer token (session ID)
        format: Bearer {session_id}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session validation result
          schema:
            $ref: '#/definitions/handlers.SessionValidationResponse'
        "400":
          description: Missing session_id in Authorization header
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error during validation
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate user session
      tags:
      - authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and session token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"